Fix task_local macro
