+++ src/prelude.rs
@@ -40,10 +40,10 @@ cfg_std! {
     pub use crate::io::prelude::WriteExt as _;
 }
 
-// cfg_default! {
-//     #[doc(no_inline)]
-//     pub use crate::task_local;
-// }
+cfg_default! {
+    #[doc(no_inline)]
+    pub use crate::task_local;
+}
 
 cfg_unstable! {
     #[doc(no_inline)]
